TODO: Add some tests

+++ src/main/java/com/example/xCode/NumbersPostResponse.java
@@ -3,11 +3,10 @@ package com.example.xCode;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 
 @Getter
 @Setter
 public class NumbersPostResponse {
-    private ArrayList<BigDecimal> numbers;
+    private ArrayList<Integer> numbers;
 }
